adjusted post model
