+++ explore.html
@@ -93,8 +93,7 @@
                 <a href="#"><img src="public/images/live-video.png" /> Watching</a>
                 <a href="#"><img src="public/images/feeling.png" /> Eating</a>
                 <a href="#"><img src="public/images/photo.png" /> Photo/Video</a>
-                <a><img src="public/images/share.png">
-                <button type="button" oneclick="post()">Post</button></a>
+                <a><img src="public/images/share.png"><button type="button" oneclick="post()">Post</button></a>
               </div>
             </div>
           </form>
